+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Changed
 - RPClient class does not use separate Thread for log processing anymore, by @HardNorth
 - Use `importlib.metadata` package for distribution data extraction for Python versions starting 3.8, by @HardNorth
+- `helpers.verify_value_length` function updated to truncate attribute keys also, by @HardNorth
 ### Removed
 - Dependency on `six`, by @HardNorth